adjahjdajJ

+++ bot.py
@@ -93,9 +93,19 @@ async def on_message(message: discord.Message):
     # 📦 LAGER KOMMANDO
     # =====================
     if content == "lager":
-        lines = ["📦 **Lagerstatus**"]
+        lager = load_lager()
+
+        if not lager:
+            await message.channel.send(
+                "📦 **Lagerstatus**\n_(Ingen varer i lageret endnu)_",
+                delete_after=6
+            )
+            return
+
+        lines =["📦 **Lagerstatus**"]
         for item, max_amount in lager.items():
-            lines.append(f"• {item}: {max_amount}")
+            lines.append(f"• **{item}**: {max_amount}")
+
         await message.channel.send("\n".join(lines), delete_after=15)
         try:
             await message.delete()
